pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = "Hosts the middleware services for IDEA-FAST, including: data transfer protocols and application consumer APIs."
authors = ["Drew Miley <drew.miley@ncl.ac.uk>", "Jay Rainey <jay.rainey@ncl.ac.uk>"]
homepage = "https://github.com/ideafast/middleware-services"
repository = "https://github.com/ideafast/middleware-services"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.12.2"
requests = "^2.24.0"
python-dotenv = "^0.14.0"
pymongo = "^3.11.2"
boto3 = "^1.16.61"
mypy-boto3-s3 = "^1.16.63"

[tool.poetry.dev-dependencies]
click = "^7.1.2"
pytest = "^6.1.2"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-aaa = "^0.11.1"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^20.11.1"
nox = "^2020.12.31"
isort = "^5.7.0"
mypy = "^0.800"

[tool.poetry.scripts]
consumer = "cli:consumer"
build = "cli:build"
publish = "cli:publish"
compose = "cli:compose"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"